authentication services

+++ src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -3,6 +3,7 @@ import { ActivatedRoute, Params, Router } from '@angular/router';
 
 import { Recipe } from '../recipe-list/recipe.model';
 import { RecipeService } from '../recipe.service';
+import { AuthService } from '../../authentication/auth.service';
 
 @Component({
 	selector: 'app-recipe-detail',
@@ -16,7 +17,8 @@ export class RecipeDetailComponent implements OnInit {
 	constructor(
 		private recipeService: RecipeService,
 		private route: ActivatedRoute,
-		private router: Router
+		private router: Router,
+		private authService: AuthService
 	) { }
 
 	ngOnInit() {
@@ -33,11 +35,15 @@ export class RecipeDetailComponent implements OnInit {
 	}
 
 	onEdit() {
-		this.router.navigate(['edit'], {relativeTo: this.route})
+		this.router.navigate(['edit'], { relativeTo: this.route })
 	}
 
 	onDelete() {
-		this.recipeService.deleteRecipe(this.id);
-		this.router.navigate(['/recipes'])
+		if (this.authService.isAuthenticated()) {
+			this.recipeService.deleteRecipe(this.id);
+			this.router.navigate(['/recipes'])
+		} else {
+			return alert('You should authenticate first!')
+		}
 	}
 }
